Remove image_utils.py
