+++ gammapy/estimators/asmooth_map.py
@@ -4,7 +4,8 @@ import numpy as np
 from astropy.convolution import Gaussian2DKernel, Tophat2DKernel
 from astropy.coordinates import Angle
 from gammapy.datasets import MapDatasetOnOff
-from gammapy.maps import WcsNDMap, scale_cube
+from gammapy.maps import WcsNDMap
+from gammapy.utils.array import scale_cube
 from gammapy.stats import significance
 
 __all__ = ["ASmoothMapEstimator"]
