+++ gammapy/makers/background/ring.py
@@ -4,7 +4,8 @@ import itertools
 import numpy as np
 from astropy.convolution import Ring2DKernel, Tophat2DKernel
 from astropy.coordinates import Angle
-from gammapy.maps import Map, scale_cube
+from gammapy.maps import Map
+from gammapy.utils.array import scale_cube
 
 __all__ = ["AdaptiveRingBackgroundMaker", "RingBackgroundMaker"]
