+++ gammapy/maps/__init__.py
@@ -6,7 +6,6 @@ from .geom import *
 from .hpx import *
 from .hpxmap import *
 from .hpxnd import *
-from .image_utils import *
 from .region import *
 from .regionnd import *
 from .wcs import *
